moving some stuff for style
